mkdocs_requirements.txt
astor==0.8.1
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
bleach==5.0.1
bracex==2.3.post1
charset-normalizer==2.1.1
click==8.1.3
csscompressor==0.9.5
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
fastjsonschema==2.16.2
ghp-import==2.1.0
gitdb==4.0.9
GitPython==3.1.27
griffe==0.22.2
htmlmin==0.1.12
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
lxml==4.9.1
Mako==1.2.3
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.3.1
mkdocs-autorefs==0.4.1
mkdocs-awesome-pages-plugin==2.8.0
mkdocs-enumerate-headings-plugin==0.5.0
mkdocs-git-authors-plugin==0.6.4
mkdocs-git-revision-date-localized-plugin==1.1.0
mkdocs-img2fig-plugin==0.9.3
mkdocs-material==8.5.3
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.5.0
mkdocs-print-site-plugin==2.3.4
mkdocs-table-reader-plugin==1.1.0
mkdocstrings==0.19.0
mkdocstrings-python==0.7.1
mkgendocs==0.9.2
mknotebooks==0.7.1
natsort==8.2.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
pipreqs==0.4.11
Pygments==2.13.0
pymdown-extensions==9.5
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tinycss2==1.1.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
watchdog==2.1.9
wcmatch==8.4.1
webencodings==0.5.1
yarg==0.1.9
zipp==3.8.1
